+++ baseapp/options.go
@@ -240,3 +240,9 @@ func (app *BaseApp) SetStreamingService(s StreamingService) {
 func (app *BaseApp) SetTxDecoder(txDecoder sdk.TxDecoder) {
 	app.txDecoder = txDecoder
 }
+
+// SetQueryMultiStore set a alternative MultiStore implementation to support grpc query service.
+// see: https://github.com/cosmos/cosmos-sdk/issues/13317
+func (app *BaseApp) SetQueryMultiStore(ms sdk.MultiStore) {
+	app.qms = ms
+}